Move categories to it's own type
