+++ client/src/app/auth/tabs/transactions/cells/editable-category-cell.tsx
@@ -1,7 +1,8 @@
-import { categories, type Transaction } from '@/types/transaction';
+import { type Transaction } from '@/types/transaction';
 import React from 'react';
 import { getCategoryLabel } from '@/lib/transactions';
 import CategoryInput from '@/components/category-input';
+import { categories } from '@/types/category';
 
 interface EditableCategoryCellProps {
   category: string;
